pyproject.toml
[project]
name = "hh-bot"
version = "0.1.0"
description = "A Telegram bot for interacting with HH.ru API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram==3.10.0",
    "aiohttp>=3.9.0,<3.10.dev0",
    "httpx==0.27.0",
    "asyncpg>=0.30.0",
    "SQLAlchemy==2.0.36",
    "alembic==1.13.3",
    "python-dotenv==1.0.1",
    "pydantic>=2.4.1,<2.9",
    "pydantic-settings==2.5.2",
    "APScheduler==3.10.4",
    "loguru==0.7.2",
    "colorama==0.4.6",
    "openai==1.52.2",
    "uvicorn==0.32.1",
    "ruff==0.6.8",
    "black==24.10.0",
    "pytest==8.3.2",
    "pytest-asyncio==0.23.8",
    "pytest-dotenv==0.5.2",
    "pyyaml>=6.0",
    "greenlet>=3.2.4",
]

[project.optional-dependencies]
dev = [
    "black==24.10.0",
    "ruff==0.6.8",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["bot*"]

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["bot", "alembic", "tests"]
respect-gitignore = true
exclude = [
    ".git",
    ".env",
    ".venv",
    "uv.lock",
    "logs",
    "alembic/versions",
]

fix = true

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "I",
    "N",
    "UP",
    "ASYNC",
    "S",
    "B",
    "A",
    "C4",
]

ignore = [
    "E501", # long lines
    "S101", # assert allowed
    "B008", # default arg call
    "A003", # allow id field
]

fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
docstring-code-format = true
skip-magic-trailing-comma = false

[tool.ruff.lint.isort]
known-first-party = ["bot"]
combine-as-imports = true

[tool.ruff.lint.per-file-ignores]
"clear_db.py" = ["S608"]
